refactor: rename templates to abies-browser and abies-browser-empty
Prepare for upcoming server template by making platform explicit in template names.

Template changes:
- Rename abies/ → abies-browser/, abies-empty/ → abies-browser-empty/
- Update shortName, identity, name, description in template.json
- Change PackageReference from Abies to Abies.Browser in both csproj
- Copy canonical abies.js with correct getAssemblyExports("Abies.Browser")

Docs updated:
- Abies.Templates/README.md, docs/getting-started/templates.md
- docs/getting-started/installation.md, docs/getting-started/project-structure.md
- docs/index.md, docs/reference/js-interop.md, docs/adr/ADR-017
- README.md, Abies.Presentation/Program.cs
