+++ Abies.Templates/templates/abies-browser-empty/.template.config/template.json
@@ -2,10 +2,10 @@
   "$schema": "http://json.schemastore.org/template",
   "author": "Maurice CGP Peters",
   "classifications": ["Web", "WebAssembly", "MVU"],
-  "identity": "Abies.Templates.Application",
-  "name": "Abies Application",
-  "shortName": "abies",
-  "description": "A minimal Abies MVU application with a counter example demonstrating the Model-View-Update pattern.",
+  "identity": "Abies.Templates.BrowserEmpty",
+  "name": "Abies Browser Application (Empty)",
+  "shortName": "abies-browser-empty",
+  "description": "An empty Abies Browser MVU application for building WebAssembly apps from scratch.",
   "sourceName": "AbiesApp",
   "defaultName": "AbiesApp",
   "preferNameDirectory": true,
