+++ Abies.Templates/templates/abies-browser-empty/AbiesApp.csproj
@@ -18,7 +18,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Abies" Version="1.0.0-*" />
+    <PackageReference Include="Abies.Browser" Version="1.0.0-*" />
   </ItemGroup>
 
 </Project>
