+++ Abies.Templates/templates/abies-browser/.template.config/template.json
@@ -2,10 +2,10 @@
   "$schema": "http://json.schemastore.org/template",
   "author": "Maurice CGP Peters",
   "classifications": ["Web", "WebAssembly", "MVU"],
-  "identity": "Abies.Templates.Empty",
-  "name": "Abies Application (Empty)",
-  "shortName": "abies-empty",
-  "description": "An empty Abies MVU application for building WebAssembly apps from scratch.",
+  "identity": "Abies.Templates.Browser",
+  "name": "Abies Browser Application",
+  "shortName": "abies-browser",
+  "description": "A minimal Abies Browser MVU application with a counter example demonstrating the Model-View-Update pattern in WebAssembly.",
   "sourceName": "AbiesApp",
   "defaultName": "AbiesApp",
   "preferNameDirectory": true,
